fix: fix input dtype

+++ fastembed/rerank/cross_encoder/onnx_text_model.py
@@ -34,12 +34,16 @@ class OnnxCrossEncoderModel(OnnxModel):
         tokenized_input = self.tokenize(query, documents, **kwargs)
 
         inputs = {
-            "input_ids": np.array([enc.ids for enc in tokenized_input]),
-            "attention_mask": np.array([enc.attention_mask for enc in tokenized_input]),
+            "input_ids": np.array([enc.ids for enc in tokenized_input], dtype=np.int64),
+            "attention_mask": np.array(
+                [enc.attention_mask for enc in tokenized_input], dtype=np.int64
+            ),
         }
         input_names = {node.name for node in self.model.get_inputs()}
         if "token_type_ids" in input_names:
-            inputs["token_type_ids"] = np.array([enc.type_ids for enc in tokenized_input])
+            inputs["token_type_ids"] = np.array(
+                [enc.type_ids for enc in tokenized_input], dtype=np.int64
+            )
 
         onnx_input = self._preprocess_onnx_input(inputs, **kwargs)
         outputs = self.model.run(self.ONNX_OUTPUT_NAMES, onnx_input)
